RSSFE-50 refactor: move tutorial for student to separate component
