+++ src/modules/TutorialPage/components/index.ts
@@ -1,3 +1,4 @@
 export { StepBlock } from './StepBlock';
 export { NoteBlock } from './NoteBlock';
 export { Switch } from './Switch';
+export { TutorialStudent } from './TutorialStudent';
